Move content of src/ewatercycle/config/_config_object.py to src/ewatercycle/config/__init__.py
Also added link from setup docs to config api docs.
